Calculable interface extended, multiply method added

+++ src/test/java/CalculatorTests.java
@@ -6,7 +6,7 @@ public class CalculatorTests {
 
     @Test
     public void matrixToStringTest(){
-        int[][] matrixData = {{1,2}, {1,1}};
+        double[][] matrixData = {{1,2}, {1,1}};
         MatrixUtil matrixUtil = new MatrixUtil(matrixData);
         MatrixUtil matrixUtilFromString = Main.getMatrixFromString("[1,2/1,1]");
